refactor(settings): move deployment time zone to general
