+++ src/Foundry/Strings/en-US/Resources.resw
@@ -810,13 +810,10 @@
   <data name="GeneralConfiguration.WinPeLanguage.Unavailable" xml:space="preserve">
     <value>No WinPE language packs were detected for the selected architecture.</value>
   </data>
-  <data name="LocalizationPage_Title.Text" xml:space="preserve">
-    <value>Localization</value>
-  </data>
-  <data name="Localization.TimeZone.Header" xml:space="preserve">
+  <data name="GeneralConfiguration.TimeZone.Header" xml:space="preserve">
     <value>Deployment time zone</value>
   </data>
-  <data name="Localization.AutomaticOption" xml:space="preserve">
+  <data name="Common.AutomaticOption" xml:space="preserve">
     <value>Automatic</value>
   </data>
   <data name="NetworkPage_Title.Text" xml:space="preserve">
@@ -1640,15 +1637,6 @@
   </data>
   <data name="Nav_HomeKey.Title" xml:space="preserve">
     <value>Home</value>
-  </data>
-  <data name="Nav_LocalizationKey.Description" xml:space="preserve">
-    <value>WinPE language and localization settings.</value>
-  </data>
-  <data name="Nav_LocalizationKey.Subtitle" xml:space="preserve">
-    <value>Localization</value>
-  </data>
-  <data name="Nav_LocalizationKey.Title" xml:space="preserve">
-    <value>Localization</value>
   </data>
   <data name="Nav_NetworkKey.Description" xml:space="preserve">
     <value>Network and connectivity configuration.</value>
