+++ src/Foundry/Strings/ja-JP/Resources.resw
@@ -810,13 +810,10 @@
   <data name="GeneralConfiguration.WinPeLanguage.Unavailable" xml:space="preserve">
     <value>選択したアーキテクチャに対して WinPE 言語パックが検出されませんでした。</value>
   </data>
-  <data name="LocalizationPage_Title.Text" xml:space="preserve">
-    <value>ローカリゼーション</value>
-  </data>
-  <data name="Localization.TimeZone.Header" xml:space="preserve">
+  <data name="GeneralConfiguration.TimeZone.Header" xml:space="preserve">
     <value>デプロイメントのタイムゾーン</value>
   </data>
-  <data name="Localization.AutomaticOption" xml:space="preserve">
+  <data name="Common.AutomaticOption" xml:space="preserve">
     <value>自動</value>
   </data>
   <data name="NetworkPage_Title.Text" xml:space="preserve">
@@ -1625,15 +1622,6 @@
   </data>
   <data name="Nav_HomeKey.Title" xml:space="preserve">
     <value>ホーム</value>
-  </data>
-  <data name="Nav_LocalizationKey.Description" xml:space="preserve">
-    <value>WinPE 言語とローカリゼーションの設定。</value>
-  </data>
-  <data name="Nav_LocalizationKey.Subtitle" xml:space="preserve">
-    <value>ローカリゼーション</value>
-  </data>
-  <data name="Nav_LocalizationKey.Title" xml:space="preserve">
-    <value>ローカリゼーション</value>
   </data>
   <data name="Nav_NetworkKey.Description" xml:space="preserve">
     <value>ネットワークと接続の構成。</value>
