--- src/Foundry/Views/LocalizationPage.xaml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Page x:Class="Foundry.Views.LocalizationPage"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      xmlns:viewModels="using:Foundry.ViewModels"
-      xmlns:wct="using:CommunityToolkit.WinUI.Controls"
-      mc:Ignorable="d">
-    <ScrollView Margin="{ThemeResource ContentPageMargin}"
-                Padding="{ThemeResource ContentPagePadding}"
-                VerticalScrollBarVisibility="Auto">
-        <StackPanel Spacing="{ThemeResource FoundrySettingsSectionSpacing}">
-            <TextBlock Text="{x:Bind ViewModel.PageTitle, Mode=OneWay}"
-                       Style="{ThemeResource FoundryPageTitleTextBlockStyle}" />
-
-            <wct:SettingsCard Header="{x:Bind ViewModel.TimeZoneHeader, Mode=OneWay}">
-                <wct:SettingsCard.HeaderIcon>
-                    <FontIcon Glyph="&#xE823;" />
-                </wct:SettingsCard.HeaderIcon>
-                <ComboBox Width="{ThemeResource FoundryWideInputMinWidth}"
-                          DisplayMemberPath="DisplayName"
-                          ItemsSource="{x:Bind ViewModel.TimeZoneOptions, Mode=OneWay}"
-                          SelectedItem="{x:Bind ViewModel.SelectedTimeZone, Mode=TwoWay}" />
-            </wct:SettingsCard>
-        </StackPanel>
-    </ScrollView>
-</Page>
