--- src/Foundry/Views/LocalizationPage.xaml.cs
@@ -1,19 +1,0 @@
-namespace Foundry.Views;
-
-public sealed partial class LocalizationPage : Page
-{
-    public LocalizationConfigurationViewModel ViewModel { get; }
-
-    public LocalizationPage()
-    {
-        ViewModel = App.GetService<LocalizationConfigurationViewModel>();
-        InitializeComponent();
-        Unloaded += OnUnloaded;
-    }
-
-    private void OnUnloaded(object sender, RoutedEventArgs e)
-    {
-        Unloaded -= OnUnloaded;
-        ViewModel.Dispose();
-    }
-}
